plot: Use `f32` instead of `f64` (#1052)
## Break Change
- Convert all `f64` parameters to `f32`.

Since the GPUI geometry module uses f32 for rendering, we are also
switching to f32. This helps avoid many as operators and unnecessary
conversions.

+++ crates/story/src/chart_story.rs
@@ -164,7 +164,7 @@ impl Render for ChartStory {
                     .child(chart_container(
                         "Pie Chart",
                         PieChart::new(self.monthly_devices.clone())
-                            .value(|d| d.desktop)
+                            .value(|d| d.desktop as f32)
                             .outer_radius(100.)
                             .color(move |d| d.color(color)),
                         true,
@@ -173,7 +173,7 @@ impl Render for ChartStory {
                     .child(chart_container(
                         "Pie Chart - Donut",
                         PieChart::new(self.monthly_devices.clone())
-                            .value(|d| d.desktop)
+                            .value(|d| d.desktop as f32)
                             .outer_radius(100.)
                             .inner_radius(60.)
                             .color(move |d| d.color(color)),
@@ -183,7 +183,7 @@ impl Render for ChartStory {
                     .child(chart_container(
                         "Pie Chart - Pad Angle",
                         PieChart::new(self.monthly_devices.clone())
-                            .value(|d| d.desktop)
+                            .value(|d| d.desktop as f32)
                             .outer_radius(100.)
                             .inner_radius(60.)
                             .pad_angle(4. / 100.)
